Remove build-queue-manager
