+++ examples/register.ts
@@ -9,9 +9,9 @@ import {
   onRenameCallbackWithCallback,
 } from "./callbacks";
 import { ItemsInfoManager, createFilesWithSize } from "./utils";
-import { VirtualDrive } from "src";
+import { QueueManager } from "./queueManager";
+import { IQueueManager, QueueItem, VirtualDrive } from "src";
 import { generateRandomFilesAndFolders } from "./utils/generate-random-file-tree";
-import { buildQueueManager } from "./build-queue-manager";
 
 console.log("Registering sync root: " + settings.syncRootPath);
 
@@ -30,9 +30,91 @@ drive.registerSyncRoot(
   settings.defaultIconPath
 );
 
-drive.connectSyncRoot();
+const handlerAdd = async (task: QueueItem) => {
+  try {
+    console.log("[EXAMPLE] File added in callback: " + task.path);
+    await new Promise((resolve) =>
+      setTimeout(() => {
+        resolve(undefined);
+      }, 1000)
+    );
+    const result = Math.random().toString(36).substring(2, 7);
+    await drive.convertToPlaceholder(task.path, result);
+    await drive.updateSyncStatus(task.path, task.isFolder, true);
+  } catch (error) {
+    console.error(error);
+  }
+};
 
-const queueManager = buildQueueManager(drive);
+const handleDehydrate = async (task: QueueItem) => {
+  try {
+    console.log("[EXAMPLE] File dehydrated in callback: " + task.path);
+    await new Promise((resolve) =>
+      setTimeout(() => {
+        resolve(undefined);
+      }, 1000)
+    );
+    console.log("Dehydrating file: " + task.path);
+    await drive.dehydrateFile(task.path);
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+const handleHydrate = async (task: QueueItem) => {
+  try {
+    console.log("[EXAMPLE] File hydrated in callback: " + task.path);
+    await new Promise((resolve) =>
+      setTimeout(() => {
+        resolve(undefined);
+      }, 1000)
+    );
+
+    const tempPath = task.path.replace(
+      settings.syncRootPath,
+      settings.serverRootPath
+    );
+
+    console.log("Hydrating file: " + task.path);
+    // await drive.transferData(tempPath, task.path);
+
+    console.log("[EXAMPLE] File trasnfer in callback: " + task.path);
+    await new Promise((resolve) =>
+      setTimeout(() => {
+        resolve(undefined);
+      }, 1000)
+    );
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+const handleChangeSize = async (task: QueueItem) => {
+  try {
+    console.log("[EXAMPLE] File size changed in callback: " + task.path);
+    await new Promise((resolve) =>
+      setTimeout(() => {
+        resolve(undefined);
+      }, 1000)
+    );
+    const result = Math.random().toString(36).substring(2, 7);
+    await drive.convertToPlaceholder(task.path, result);
+    await drive.updateFileIdentity(task.path, result, false);
+    await drive.updateSyncStatus(task.path, task.isFolder, true);
+    // await drive.updateFileSize(task.path);
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+const queueManager: IQueueManager = new QueueManager({
+  handleAdd: handlerAdd,
+  handleHydrate: handleHydrate,
+  handleDehydrate: handleDehydrate,
+  handleChangeSize: handleChangeSize,
+});
+
+drive.connectSyncRoot();
 
 const fileGenerationOptions = {
   rootPath: '',
